consolidate IAeadBlockCipherModeTransform tests into shared abstract base
Move duplicated nonce/AAD/plaintext sensitivity, determinism, empty-AAD
equivalence, and AAD-after-encrypt/decrypt lifecycle tests from each
concrete AEAD mode test class into AeadBlockCipherModeTests<TTransform>
partials so every IAeadBlockCipherModeTransform implementation
(GCM, CCM, EAX, OCB, SIV, GCM-SIV) is exercised against the same
contract. Add a NonceAffectsCiphertext virtual property, overridden
to false in SivModeTransformTests where the supplied IV is ignored.
