+++ Bodu.Security.Cryptography/test/Security.Cryptography/AeadBlockCipherModeTests.Decrypt.cs
@@ -105,6 +105,37 @@ public abstract partial class AeadBlockCipherModeTests<TTransform>
     }
 
     /// <summary>
+    /// Verifies that decrypting under a different nonce than was used to produce the
+    /// ciphertext causes <see cref="IAeadBlockCipherModeTransform.Decrypt" /> to throw
+    /// <see cref="CryptographicException" />. Skipped for modes whose ciphertext does not
+    /// depend on the supplied IV.
+    /// </summary>
+    [TestMethod]
+    public void Decrypt_WithWrongNonce_ShouldThrowCryptographicException()
+    {
+        if (!NonceAffectsCiphertext)
+        {
+            Assert.Inconclusive($"{typeof(TTransform).Name} ciphertext does not depend on the supplied IV.");
+            return;
+        }
+
+        var cipher = new MonitoringBlockCipher(ExpectedBlockSize, xorMask: 0xAA);
+        var ivEnc = CreateInitializationVector();
+        var ivDec = CreateInitializationVector();
+        ivDec[0] = 0x01; // nonce differs
+        var plaintext = new byte[ExpectedBlockSize];
+
+        var encTransform = CreateTransform(cipher, ivEnc);
+        var buf = new byte[plaintext.Length + encTransform.TagSize];
+        encTransform.Encrypt(plaintext, buf);
+
+        var decTransform = CreateTransform(cipher, ivDec);
+        Assert.ThrowsExactly<CryptographicException>(() =>
+            decTransform.Decrypt(buf, new byte[plaintext.Length]),
+            "Decrypting with a different nonce must throw CryptographicException.");
+    }
+
+    /// <summary>
     /// Verifies that supplying different associated data to the decrypting instance than was
     /// used during encryption causes <see cref="IAeadBlockCipherModeTransform.Decrypt" /> to
     /// throw <see cref="CryptographicException" />, confirming the AAD is bound into the tag.
